Implement game() and delete displayRound()
Previously, displayRound() was a separate function that displayed the
results of single round. Because it contained the same logic necessary
to update scores in game(), it was deleted and combined into a single
if-else block for displaying and updating scores.

The remainder of game() was implemented by calling previously
defined helper functions.